Refactor: move Cursor to `src` module; add type hints
