+++ Launcher.py
@@ -9,6 +9,7 @@ from typing import ClassVar
 import pygame
 
 from src.constants import COLORS, PENETRATION_COLORS
+from src.cursor import Cursor
 
 if getattr(sys, 'frozen', False) and hasattr(sys, '_MEIPASS'):
     BASE_DIR = Path(sys._MEIPASS)
@@ -786,40 +787,6 @@ class HealthBar(pygame.sprite.Sprite):
             camera.apply(outline_rect),
             1,
         )
-
-
-class Cursor(pygame.sprite.Sprite):
-    """Implements gunsight/cursor."""
-
-    OUTER_RADIUS: ClassVar = 20
-    INNER_RADIUS: ClassVar = 2
-
-    def __init__(self):
-        super().__init__()
-        self.image = pygame.Surface(
-            size=(2 * self.OUTER_RADIUS, 2 * self.OUTER_RADIUS),
-            flags=pygame.SRCALPHA,
-        )
-        self.rect = self.image.get_rect()
-        center = self.rect.center
-        pygame.draw.circle(
-            surface=self.image,
-            color=COLORS['WHITE'],
-            center=center,
-            radius=self.OUTER_RADIUS,
-            width=2,
-        )
-        pygame.draw.circle(
-            surface=self.image,
-            color=COLORS['RED'],
-            center=center,
-            radius=self.INNER_RADIUS,
-        )
-
-    def draw(self, *, surface, center_pos):
-        """Blit to `surface`, centered on `center_pos`."""
-
-        surface.blit(self.image, (center_pos[0] - self.OUTER_RADIUS, center_pos[1] - self.OUTER_RADIUS))
 
 
 def manhattan_distance(a, b):
